requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
confection==0.1.0
cymem==2.0.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
Flask
frozenlist
googlesearch-python==1.2.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
langcodes==3.3.0
MarkupSafe==2.1.3
multidict==6.0.4
murmurhash==1.0.9
numpy
openai==0.27.8
packaging==23.1
pathy==0.10.2
preshed==3.0.8
pydantic==1.10.11
requests
scikit-learn
scipy
smart-open
soupsieve
spacy
spacy-legacy
spacy-loggers
srsly
thinc
threadpoolctl
tqdm
typer
typing-extensions
urllib3
wikipedia
wasabi
Werkzeug
yarl
Flask-SQLAlchemy
PyMySQL
